Split out more commands and removed the duplicated ident.c file.
