+++ .gitignore
@@ -1,3 +1,4 @@
 .settings
 .cproject
 .project
+src/objects
