+++ src/amnuts230.c
@@ -6109,103 +6109,6 @@ exec_com(UR_OBJECT user, char *inpstr, enum cmd_value defaultcmd)
 }
 
 
-/*
- * Display details of room
- */
-void
-look(UR_OBJECT user)
-{
-  char temp[81];
-  RM_OBJECT rm;
-  UR_OBJECT u;
-  int i, exits, users;
-
-  rm = user->room;
-
-  if (is_private_room(rm)) {
-    vwrite_user(user, "\n~FCRoom: ~FR%s\n\n",
-                is_personal_room(rm) ? rm->show_name : rm->name);
-  } else {
-    vwrite_user(user, "\n~FCRoom: ~FG%s\n\n",
-                is_personal_room(rm) ? rm->show_name : rm->name);
-  }
-  if (user->show_rdesc) {
-    write_user(user, user->room->desc);
-  }
-  exits = 0;
-  strcpy(text, "\n~FCExits are:");
-  for (i = 0; i < MAX_LINKS; ++i) {
-    if (!rm->link[i]) {
-      break;
-    }
-    if (is_private_room(rm->link[i])) {
-      sprintf(temp, "  ~FR%s", rm->link[i]->name);
-    } else {
-      sprintf(temp, "  ~FG%s", rm->link[i]->name);
-    }
-    strcat(text, temp);
-    ++exits;
-  }
-#ifdef NETLINKS
-  if (rm->netlink && rm->netlink->stage == UP) {
-    if (rm->netlink->allow == IN) {
-      sprintf(temp, "  ~FR%s*", rm->netlink->service);
-    } else {
-      sprintf(temp, "  ~FG%s*", rm->netlink->service);
-    }
-    strcat(text, temp);
-  } else
-#endif
-  if (!exits) {
-    strcpy(text, "\n~FCThere are no exits.");
-  }
-  strcat(text, "\n\n");
-  write_user(user, text);
-  users = 0;
-  for (u = user_first; u; u = u->next) {
-    if (u->room != rm || u == user || (!u->vis && u->level > user->level)) {
-      continue;
-    }
-    if (!users++) {
-      write_user(user, "~FG~OLYou can see:\n");
-    }
-    vwrite_user(user, "     %s%s~RS %s~RS  %s\n", u->vis ? " " : "~FR*~RS",
-                u->recap, u->desc, u->afk ? "~BR(AFK)" : "");
-  }
-  if (!users) {
-    write_user(user, "~FGYou are all alone here.\n");
-  }
-  write_user(user, "\n");
-  strcpy(text, "Access is ");
-  if (is_personal_room(rm)) {
-    strcat(text, "personal ");
-    if (is_private_room(rm)) {
-      strcat(text, "~FR(locked)~RS");
-    } else {
-      strcat(text, "~FG(unlocked)~RS");
-    }
-  } else {
-    if (is_fixed_room(rm)) {
-      strcat(text, "~FRfixed~RS to ");
-    } else {
-      strcat(text, "set to ");
-    }
-    if (is_private_room(rm)) {
-      strcat(text, "~FRPRIVATE~RS");
-    } else {
-      strcat(text, "~FGPUBLIC~RS");
-    }
-  }
-  sprintf(temp, " and there %s ~OL~FM%d~RS message%s on the board.\n",
-          PLTEXT_IS(rm->mesg_cnt), rm->mesg_cnt, PLTEXT_S(rm->mesg_cnt));
-  strcat(text, temp);
-  write_user(user, text);
-  if (*rm->topic) {
-    vwrite_user(user, "~FG~OLCurrent topic:~RS %s\n", rm->topic);
-  } else {
-    write_user(user, "~FGNo topic has been set yet.\n");
-  }
-}
 
 
 
@@ -6588,322 +6491,9 @@ help_amnuts_credits(UR_OBJECT user)
 }
 
 
-/*
- * Show some user stats
- */
-void
-status(UR_OBJECT user)
-{
-  char ir[ROOM_NAME_LEN + 1], text2[ARR_SIZE], text3[ARR_SIZE], rm[4],
-    qcall[USER_NAME_LEN];
-  char email[82], nm[5], muzlev[20], arrlev[20];
-  UR_OBJECT u;
-  int days, hours, mins, hs, on, cnt, newmail;
-  time_t t_expire;
-
-  if (word_count < 2) {
-    u = user;
-    on = 1;
-  } else {
-    u = retrieve_user(user, word[1]);
-    if (!u) {
-      return;
-    }
-    on = retrieve_user_type == 1;
-  }
-  write_user(user, "\n\n");
-  if (!on) {
-    write_user(user,
-               "+----- ~FCUser Info~RS -- ~FB(not currently logged on)~RS -------------------------------+\n");
-  } else {
-    write_user(user,
-               "+----- ~FCUser Info~RS -- ~OL~FY(currently logged on)~RS -----------------------------------+\n");
-  }
-  sprintf(text2, "%s~RS %s", u->recap, u->desc);
-  cnt = 45 + teslen(text2, 45);
-  vwrite_user(user, "Name   : %-*.*s~RS Level : %s\n", cnt, cnt, text2,
-              user_level[u->level].name);
-  mins = (int) (time(0) - u->last_login) / 60;
-  days = u->total_login / 86400;
-  hours = (u->total_login % 86400) / 3600;
-  if (!u->invite_room) {
-    strcpy(ir, "<nowhere>");
-  } else {
-    strcpy(ir, u->invite_room->name);
-  }
-  if (!u->age) {
-    sprintf(text2, "Unknown");
-  } else {
-    sprintf(text2, "%d", u->age);
-  }
-  if (!on || u->login) {
-    *text3 = '\0';
-  } else {
-    sprintf(text3, "            Online for : %d min%s", mins, PLTEXT_S(mins));
-  }
-  vwrite_user(user, "Gender : %-8s      Age : %-8s %s\n", sex[u->gender],
-              text2, text3);
-  if (!*u->email) {
-    strcpy(email, "Currently unset");
-  } else {
-    strcpy(email, u->email);
-    if (u->mail_verified) {
-      strcat(email, " ~FB~OL(VERIFIED)~RS");
-    }
-    if (u->hideemail) {
-      if (user->level >= WIZ || u == user) {
-        strcat(email, " ~FB~OL(HIDDEN)~RS");
-      } else {
-        strcpy(email, "Currently only on view to the Wizzes");
-      }
-    }
-  }
-  vwrite_user(user, "Email Address : %s\n", email);
-  vwrite_user(user, "Homepage URL  : %s\n",
-              !*u->homepage ? "Currently unset" : u->homepage);
-  vwrite_user(user, "ICQ Number    : %s\n",
-              !*u->icq ? "Currently unset" : u->icq);
-  mins = (u->total_login % 3600) / 60;
-  vwrite_user(user,
-              "Total Logins  : %-9d  Total login : %d day%s, %d hour%s, %d minute%s\n",
-              u->logons, days, PLTEXT_S(days), hours, PLTEXT_S(hours), mins,
-              PLTEXT_S(mins));
-  write_user(user,
-             "+----- ~FCGeneral Info~RS ---------------------------------------------------------+\n");
-  vwrite_user(user, "Enter Msg     : %s~RS %s\n", u->recap, u->in_phrase);
-  vwrite_user(user, "Exit Msg      : %s~RS %s~RS to the...\n", u->recap,
-              u->out_phrase);
-  newmail = mail_sizes(u->name, 1);
-  if (!newmail) {
-    sprintf(nm, "NO");
-  } else {
-    sprintf(nm, "%d", newmail);
-  }
-  /* FIXME: Use sentinel other JAILED */
-  if (!on || u->login) {
-    vwrite_user(user, "New Mail      : %-13.13s  Muzzled : %-13.13s\n", nm,
-                noyes[(u->muzzled != JAILED)]);
-  } else {
-    vwrite_user(user,
-                "Invited to    : %-13.13s  Muzzled : %-13.13s  Ignoring : %-13.13s\n",
-                ir, noyes[(u->muzzled != JAILED)], noyes[u->ignall]);
-#ifdef NETLINKS
-    if (u->type == REMOTE_TYPE || !u->room) {
-      hs = 0;
-      sprintf(ir, "<off site>");
-    } else {
-#endif
-      hs = 1;
-      sprintf(ir, "%s", u->room->name);
-#ifdef NETLINKS
-    }
-#endif
-    vwrite_user(user,
-                "In Area       : %-13.13s  At home : %-13.13s  New Mail : %-13.13s\n",
-                ir, noyes[hs], nm);
-  }
-  vwrite_user(user,
-              "Killed %d people, and died %d times.  Energy : %d, Bullets : %d\n",
-              u->kills, u->deaths, u->hps, u->bullets);
-  if (u == user || user->level >= WIZ) {
-    write_user(user,
-               "+----- ~FCUser Only Info~RS -------------------------------------------------------+\n");
-    vwrite_user(user,
-                "Char echo     : %-13.13s  Wrap    : %-13.13s  Monitor  : %-13.13s\n",
-                noyes[u->charmode_echo], noyes[u->wrap], noyes[u->monitor]);
-    if (u->lroom == 2) {
-      strcpy(rm, "YES");
-    } else {
-      strcpy(rm, noyes[u->lroom]);
-    }
-    vwrite_user(user,
-                "Colours       : %-13.13s  Pager   : %-13d  Logon rm : %-13.13s\n",
-                noyes[u->colour], u->pager, rm);
-    if (!*u->call) {
-      strcpy(qcall, "<no one>");
-    } else {
-      strcpy(qcall, u->call);
-    }
-    vwrite_user(user,
-                "Quick call to : %-13.13s  Autofwd : %-13.13s  Verified : %-13.13s\n",
-                qcall, noyes[u->autofwd], noyes[u->mail_verified]);
-    if (on && !u->login) {
-      if (u == user && user->level < WIZ) {
-        vwrite_user(user, "On from site  : %s\n", u->site);
-      } else {
-        vwrite_user(user, "On from site  : %-42.42s  Port : %s\n", u->site,
-                    u->site_port);
-      }
-    }
-  }
-  if (user->level >= WIZ) {
-    write_user(user,
-               "+----- ~OL~FCWiz Only Info~RS --------------------------------------------------------+\n");
-    /* FIXME: Use sentinel other JAILED */
-    if (u->muzzled == JAILED) {
-      strcpy(muzlev, "Unmuzzled");
-    } else {
-      strcpy(muzlev, user_level[u->muzzled].name);
-    }
-    /* FIXME: Use sentinel other JAILED */
-    if (u->arrestby == JAILED) {
-      strcpy(arrlev, "Unarrested");
-    } else {
-      strcpy(arrlev, user_level[u->arrestby].name);
-    }
-    vwrite_user(user,
-                "Unarrest Lev  : %-13.13s  Arr lev : %-13.13s  Muz Lev  : %-13.13s\n",
-                user_level[u->unarrest].name, arrlev, muzlev);
-    if (u->lroom == 2) {
-      sprintf(rm, "YES");
-    } else {
-      sprintf(rm, "NO");
-    }
-    vwrite_user(user, "Logon room    : %-38.38s  Shackled : %s\n",
-                u->logout_room, rm);
-    vwrite_user(user, "Last site     : %s\n", u->last_site);
-    t_expire =
-      u->last_login + 86400 * (u->level ==
-                               NEW ? NEWBIE_EXPIRES : USER_EXPIRES);
-    strftime(text2, ARR_SIZE, "%a %Y-%m-%d %H:%M:%S", localtime(&t_expire));
-    vwrite_user(user, "User Expires  : %-13.13s  On date : %s\n",
-                noyes[u->expire], text2);
-  }
-  write_user(user,
-             "+----------------------------------------------------------------------------+\n\n");
-  if (u != user) {
-    done_retrieve(u);
-  }
-}
 
 
-/*
- * Examine a user
- */
-void
-examine(UR_OBJECT user)
-{
-  char filename[80], text2[ARR_SIZE];
-  FILE *fp;
-  UR_OBJECT u;
-  int on, days, hours, mins, timelen, days2, hours2, mins2, idle, cnt,
-    newmail;
-  int lastdays, lasthours, lastmins;
 
-  if (word_count < 2) {
-    u = user;
-    on = 1;
-  } else {
-    u = retrieve_user(user, word[1]);
-    if (!u) {
-      return;
-    }
-    on = retrieve_user_type == 1;
-  }
-
-  days = u->total_login / 86400;
-  hours = (u->total_login % 86400) / 3600;
-  mins = (u->total_login % 3600) / 60;
-  timelen = (int) (time(0) - u->last_login);
-  days2 = timelen / 86400;
-  hours2 = (timelen % 86400) / 3600;
-  mins2 = (timelen % 3600) / 60;
-
-  write_user(user,
-             "+----------------------------------------------------------------------------+\n");
-  sprintf(text2, "%s~RS %s", u->recap, u->desc);
-  cnt = 45 + teslen(text2, 45);
-  vwrite_user(user, "Name   : %-*.*s~RS Level : %s\n", cnt, cnt, text2,
-              user_level[u->level].name);
-  if (!on) {
-    lastdays = u->last_login_len / 86400;
-    lasthours = (u->last_login_len % 86400) / 3600;
-    lastmins = (u->last_login_len % 3600) / 60;
-
-    strftime(text2, ARR_SIZE, "%a %Y-%m-%d %H:%M:%S",
-             localtime(&u->last_login));
-    vwrite_user(user, "Last login : %s\n", text2);
-    vwrite_user(user, "Which was  : %d day%s, %d hour%s, %d minute%s ago\n",
-                days2, PLTEXT_S(days2), hours2, PLTEXT_S(hours2), mins2,
-                PLTEXT_S(mins2));
-    vwrite_user(user,
-                "Was on for : %d day%s, %d hour%s, %d minute%s\nTotal login: %d day%s, %d hour%s, %d minute%s\n",
-                lastdays, PLTEXT_S(lastdays), lasthours, PLTEXT_S(lasthours),
-                lastmins, PLTEXT_S(lastmins), days, PLTEXT_S(days), hours,
-                PLTEXT_S(hours), mins, PLTEXT_S(mins));
-    if (user->level >= WIZ) {
-      vwrite_user(user, "Last site  : %s\n", u->last_site);
-    }
-    newmail = mail_sizes(u->name, 1);
-    if (newmail) {
-      vwrite_user(user, "%s~RS has unread mail (%d).\n", u->recap, newmail);
-    }
-    write_user(user,
-               "+----- ~OL~FCProfile~RS --------------------------------------------------------------+\n\n");
-    sprintf(filename, "%s/%s/%s.P", USERFILES, USERPROFILES, u->name);
-    fp = fopen(filename, "r");
-    if (!fp) {
-      write_user(user, "User has not yet witten a profile.\n\n");
-    } else {
-      fclose(fp);
-      more(user, user->socket, filename);
-    }
-    write_user(user,
-               "+----------------------------------------------------------------------------+\n\n");
-    done_retrieve(u);
-    return;
-  }
-  idle = (int) (time(0) - u->last_input) / 60;
-  if (u->malloc_start) {
-    vwrite_user(user, "Ignoring all: ~FCUsing Line Editor\n");
-  } else {
-    vwrite_user(user, "Ignoring all: %s\n", noyes[u->ignall]);
-  }
-  strftime(text2, ARR_SIZE, "%a %Y-%m-%d %H:%M:%S",
-           localtime(&u->last_login));
-  vwrite_user(user,
-              "On since    : %s\nOn for      : %d day%s, %d hour%s, %d minute%s\n",
-              text2, days2, PLTEXT_S(days2), hours2, PLTEXT_S(hours2), mins2,
-              PLTEXT_S(mins2));
-  if (u->afk) {
-    vwrite_user(user, "Idle for    : %d minute%s ~BR(AFK)\n", idle,
-                PLTEXT_S(idle));
-    if (*u->afk_mesg) {
-      vwrite_user(user, "AFK message : %s\n", u->afk_mesg);
-    }
-  } else {
-    vwrite_user(user, "Idle for    : %d minute%s\n", idle, PLTEXT_S(idle));
-  }
-  vwrite_user(user, "Total login : %d day%s, %d hour%s, %d minute%s\n", days,
-              PLTEXT_S(days), hours, PLTEXT_S(hours), mins, PLTEXT_S(mins));
-  if (u->socket >= 0) {
-    if (user->level >= WIZ) {
-      vwrite_user(user, "Site        : %-40.40s  Port : %s\n", u->site,
-                  u->site_port);
-    }
-#ifdef NETLINKS
-    if (!u->room) {
-      vwrite_user(user, "Home service: %s\n", u->netlink->service);
-    }
-#endif
-  }
-  newmail = mail_sizes(u->name, 1);
-  if (newmail) {
-    vwrite_user(user, "%s~RS has unread mail (%d).\n", u->recap, newmail);
-  }
-  write_user(user,
-             "+----- ~OL~FCProfile~RS --------------------------------------------------------------+\n\n");
-  sprintf(filename, "%s/%s/%s.P", USERFILES, USERPROFILES, u->name);
-  fp = fopen(filename, "r");
-  if (!fp) {
-    write_user(user, "User has not yet written a profile.\n\n");
-  } else {
-    fclose(fp);
-    more(user, user->socket, filename);
-  }
-  write_user(user,
-             "+----------------------------------------------------------------------------+\n\n");
-}
 
 
 /*
@@ -7400,36 +6990,7 @@ prompt(UR_OBJECT user)
 }
 
 
-/*
- * Switch prompt on and off
- */
-void
-toggle_prompt(UR_OBJECT user)
-{
-  if (user->prompt) {
-    write_user(user, "Prompt ~FROFF.\n");
-    user->prompt = 0;
-    return;
-  }
-  write_user(user, "Prompt ~FGON.\n");
-  user->prompt = 1;
-}
 
-
-/*
- * Switch between command and speech mode
- */
-void
-toggle_mode(UR_OBJECT user)
-{
-  if (user->command_mode) {
-    write_user(user, "Now in ~FGSPEECH~RS mode.\n");
-    user->command_mode = 0;
-    return;
-  }
-  write_user(user, "Now in ~FCCOMMAND~RS mode.\n");
-  user->command_mode = 1;
-}
 
 
 /*
@@ -7453,95 +7014,10 @@ toggle_charecho(UR_OBJECT user)
 }
 
 
-/*
- * Set user description
- */
-void
-set_desc(UR_OBJECT user, char *inpstr)
-{
-  if (word_count < 2) {
-    vwrite_user(user, "Your current description is: %s\n", user->desc);
-    return;
-  }
-  if (strstr(colour_com_strip(inpstr), "(CLONE)")) {
-    write_user(user, "You cannot have that description.\n");
-    return;
-  }
-  if (strlen(inpstr) > USER_DESC_LEN) {
-    write_user(user, "Description too long.\n");
-    return;
-  }
-  strcpy(user->desc, inpstr);
-  write_user(user, "Description set.\n");
-  /* check to see if user should be promoted */
-  check_autopromote(user, 2);
-}
 
 
-/*
- * Set in and out phrases
- */
-void
-set_iophrase(UR_OBJECT user, char *inpstr)
-{
-  if (strlen(inpstr) > PHRASE_LEN) {
-    write_user(user, "Phrase too long.\n");
-    return;
-  }
-  if (com_num == INPHRASE) {
-    if (word_count < 2) {
-      vwrite_user(user, "Your current in phrase is: %s\n", user->in_phrase);
-      return;
-    }
-    strcpy(user->in_phrase, inpstr);
-    write_user(user, "In phrase set.\n");
-    return;
-  }
-  if (word_count < 2) {
-    vwrite_user(user, "Your current out phrase is: %s\n", user->out_phrase);
-    return;
-  }
-  strcpy(user->out_phrase, inpstr);
-  write_user(user, "Out phrase set.\n");
-}
 
 
-/*
- * Enter user profile
- */
-void
-enter_profile(UR_OBJECT user, char *inpstr)
-{
-  char filename[80];
-  char *c;
-  FILE *fp;
-
-  if (inpstr) {
-    if (word_count < 2) {
-      write_user(user, "\n~BB*** Writing profile ***\n\n");
-      user->misc_op = 5;
-      editor(user, NULL);
-      return;
-    }
-    strcat(inpstr, "\n");
-  } else {
-    inpstr = user->malloc_start;
-  }
-  sprintf(filename, "%s/%s/%s.P", USERFILES, USERPROFILES, user->name);
-  fp = fopen(filename, "w");
-  if (!fp) {
-    vwrite_user(user, "%s: cannot save your profile.\n", syserror);
-    write_syslog(SYSLOG | ERRLOG, 0,
-                 "ERROR: Cannot open file %s to write in enter_profile().\n",
-                 filename);
-    return;
-  }
-  for (c = inpstr; *c; ++c) {
-    putc(*c, fp);
-  }
-  fclose(fp);
-  write_user(user, "Profile stored.\n");
-}
 
 /*
  * Get user macros
@@ -7936,34 +7412,7 @@ show_ignlist(UR_OBJECT user)
 }
 
 
-/*
- * Switch ignoring all on and off
- */
-void
-toggle_ignall(UR_OBJECT user)
-{
-  if (!user->ignall) {
-    write_user(user, "You are now ignoring everyone.\n");
-    if (user->vis) {
-      vwrite_room_except(user->room, user,
-                         "%s~RS is now ignoring everyone.\n", user->recap);
-    } else {
-      vwrite_room_except(user->room, user,
-                         "%s~RS is now ignoring everyone.\n", invisname);
-    }
-    user->ignall = 1;
-    return;
-  }
-  write_user(user, "You will now hear everyone again.\n");
-  if (user->vis) {
-    vwrite_room_except(user->room, user, "%s~RS is listening again.\n",
-                       user->recap);
-  } else {
-    vwrite_room_except(user->room, user, "%s~RS is listening again.\n",
-                       invisname);
-  }
-  user->ignall = 0;
-}
+
 
 
 /*
@@ -8331,73 +7780,6 @@ get_time(UR_OBJECT user)
 }
 
 
-/*
- * Show version number and some small stats of the talker
- */
-void
-show_version(UR_OBJECT user)
-{
-  int rms;
-  RM_OBJECT rm;
-
-  write_user(user,
-             "+----------------------------------------------------------------------------+\n");
-  write_user(user, align_string(1, 78, 1, "|", "~OL%s~RS", TALKER_NAME));
-  write_user(user,
-             "+----------------------------------------------------------------------------+\n");
-  vwrite_user(user,
-              "| Logons this current boot : ~OL%4d~RS new users, ~OL%4d~RS old users~RS                  |\n",
-              amsys->logons_new, amsys->logons_old);
-  vwrite_user(user,
-              "| Total number of users    : ~OL%-4d~RS  Maximum online users     : ~OL%-3d~RS            |\n",
-              amsys->user_count, amsys->max_users);
-  rms = 0;
-  for (rm = room_first; rm; rm = rm->next) {
-    ++rms;
-  }
-  vwrite_user(user,
-              "| Total number of rooms    : ~OL%-3d~RS   Swear ban currently on   : ~OL%s~RS            |\n",
-              rms, minmax[amsys->ban_swearing]);
-  vwrite_user(user,
-              "| Smail auto-forwarding on : ~OL%-3s~RS   Auto purge on            : ~OL%-3s~RS            |\n",
-              noyes[amsys->forwarding], noyes[amsys->auto_purge_date != -1]);
-  vwrite_user(user,
-              "| Maximum smail copies     : ~OL%-3d~RS   Names can be recapped    : ~OL%-3s~RS            |\n",
-              MAX_COPIES, noyes[amsys->allow_recaps]);
-  vwrite_user(user,
-              "| Personal rooms active    : ~OL%-3s~RS   Maximum user idle time   : ~OL%-3d~RS mins~RS       |\n",
-              noyes[amsys->personal_rooms], amsys->user_idle_time / 60);
-  if (user->level >= WIZ) {
-#ifdef NETLINKS
-    write_user(user,
-               "| Compiled netlinks        : ~OLYES~RS                                             |\n");
-#else
-    write_user(user,
-               "| Compiled netlinks        : ~OLNO~RS                                              |\n");
-#endif
-  }
-  /* YOU MUST *NOT* ALTER THE REMAINING LINES */
-  write_user(user,
-             "+----------------------------------------------------------------------------+\n");
-  vwrite_user(user,
-              "| ~FC~OLAmnuts version %-21.21s (C) Andrew Collington, September 2001~RS |\n",
-              AMNUTSVER);
-  vwrite_user(user,
-              "| ~FC~OLBuilt %s %s~RS                                                 |\n",
-              __DATE__, __TIME__);
-  write_user(user,
-             "+----------------------------------------------------------------------------+\n");
-  write_user(user,
-             "| Running Ardant's Ident Daemon (ArIdent) code, version 2.0.2                |\n");
-  write_user(user,
-             "| Running Ardant's Universal Pager code                                      |\n");
-  write_user(user,
-             "| Running Arny's Seamless reboot code (based on phypor's EWTOO sreboot)      |\n");
-  write_user(user,
-             "| Running Silver's spodlist code (converted from PG+ spodlist)               |\n");
-  write_user(user,
-             "+----------------------------------------------------------------------------+\n");
-}
