+++ src/includes/rooms.h
@@ -1,0 +1,21 @@
+#ifndef AMNUTS_ROOMS_H
+#define AMNUTS_ROOMS_H
+
+struct priv_room_struct
+{
+  const char *name;
+  enum lvl_value level;
+};
+
+/*
+ * The rooms listed here are just examples of what can be added
+ * You may add more or remove as many as you like, but you MUST
+ * keep the stopping clause in
+ */
+static const struct priv_room_struct priv_room[] = {
+  {"wizroom", WIZ},             /* a room for wizzes+ only */
+  {"andys_computer", GOD},      /* only top people can get in this place! */
+  {NULL, NUM_LEVELS}            /* stopping clause */
+};
+
+#endif
