+++ src/messages.c
@@ -352,124 +352,9 @@ send_mail(UR_OBJECT user, char *to, char *ptr, int iscopy)
 }
 
 
-/*
- * Send mail message
- */
-void
-smail(UR_OBJECT user, char *inpstr)
-{
-  if (inpstr) {
-    static const char usage[] = "Usage: smail <user>[@<service>] [<text>]\n";
-    char *remote;
-
-    /* FIXME: Use sentinel other JAILED */
-    if (user->muzzled != JAILED) {
-      write_user(user, "You are muzzled, you cannot mail anyone.\n");
-      return;
-    }
-    if (word_count < 2) {
-      write_user(user, usage);
-      return;
-    }
-    *word[1] = toupper(*word[1]);
-    remote = strchr(word[1], '@');
-    /* See if user exists */
-    if (!remote) {
-      UR_OBJECT u = get_user(word[1]);
-
-      /* See if user has local account */
-      if (!find_user_listed(word[1])) {
-        if (!u) {
-          write_user(user, nosuchuser);
-        } else {
-          vwrite_user(user,
-                      "%s is a remote user and does not have a local account.\n",
-                      u->name);
-        }
-        return;
-      }
-      if (u) {
-        if (u == user && user->level < ARCH) {
-          write_user(user,
-                     "Trying to mail yourself is the fifth sign of madness.\n");
-          return;
-        }
-        /* FIXME: Should check for offline users as well */
-        if (check_igusers(u, user) && user->level < GOD) {
-          vwrite_user(user, "%s~RS is ignoring smails from you.\n", u->recap);
-          return;
-        }
-        strcpy(word[1], u->name);
-      }
-    } else if (remote == word[1]) {
-      write_user(user, "Users name missing before @ sign.\n");
-      return;
-    }
-    strcpy(user->mail_to, word[1]);
-    if (word_count < 3) {
-#ifdef NETLINKS
-      if (user->type == REMOTE_TYPE) {
-        write_user(user,
-                   "Sorry, due to software limitations remote users cannot use the line editor.\nUse the \".smail <user> <text>\" method instead.\n");
-        return;
-      }
-#endif
-      vwrite_user(user, "\n~BB*** Writing mail message to %s ***\n\n",
-                  user->mail_to);
-      user->misc_op = 4;
-      editor(user, NULL);
-      return;
-    }
-    /* One line mail */
-    strcat(inpstr, "\n");
-    inpstr = remove_first(inpstr);
-  } else {
-    if (*user->malloc_end-- != '\n') {
-      *user->malloc_end-- = '\n';
-    }
-    inpstr = user->malloc_start;
-  }
-  strtoname(user->mail_to);
-  send_mail(user, user->mail_to, inpstr, 0);
-  send_copies(user, inpstr);
-  *user->mail_to = '\0';
-}
 
 
-/*
- * Read your mail
- */
-void
-rmail(UR_OBJECT user)
-{
-  char filename[80];
-  int ret;
 
-  sprintf(filename, "%s/%s/%s.M", USERFILES, USERMAILS, user->name);
-
-  /* Just reading the one message or all new mail */
-  if (word_count > 1) {
-    strtolower(word[1]);
-    if (!strcmp(word[1], "new")) {
-      read_new_mail(user);
-    } else {
-      read_specific_mail(user);
-    }
-    return;
-  }
-  /* Update last read / new mail received time at head of file */
-  if (!reset_mail_counts(user)) {
-    write_user(user, "You do not have any mail.\n");
-    return;
-  }
-  /* Reading the whole mail box */
-  write_user(user, "\n~BB*** Your mailbox has the following messages ***\n\n");
-  user->filepos = 0;
-  ret = more(user, user->socket, filename);
-  if (ret == 1) {
-    user->misc_op = 2;
-  }
-}
 
 
 /*
@@ -596,95 +481,9 @@ read_new_mail(UR_OBJECT user)
 }
 
 
-/*
- * Delete some or all of your mail. A problem here is once we have deleted
- * some mail from the file do we mark the file as read? If not we could
- * have a situation where the user deletes all his mail but still gets
- * the YOU HAVE UNREAD MAIL message on logging on if the idiot forgot to
- * read it first.
- */
-void
-dmail(UR_OBJECT user)
-{
-  int num, cnt;
-  char filename[80];
-
-  if (word_count < 2) {
-    write_user(user, "Usage: dmail all\n");
-    write_user(user, "       dmail <#>\n");
-    write_user(user, "       dmail to <#>\n");
-    write_user(user, "       dmail from <#> to <#>\n");
-    return;
-  }
-  if (get_wipe_parameters(user) == -1) {
-    return;
-  }
-  num = mail_sizes(user->name, 0);
-  if (!num) {
-    write_user(user, "You have no mail to delete.\n");
-    return;
-  }
-  sprintf(filename, "%s/%s/%s.M", USERFILES, USERMAILS, user->name);
-  if (user->wipe_from == -1) {
-    write_user(user, "\07~OL~FR~LIDelete all of your mail?~RS (y/n): ");
-    user->misc_op = 18;
-    return;
-  }
-  if (user->wipe_from > num) {
-    vwrite_user(user, "You only have %d mail message%s.\n", num,
-                PLTEXT_S(num));
-    return;
-  }
-  cnt = wipe_messages(filename, user->wipe_from, user->wipe_to, 1);
-  reset_mail_counts(user);
-  if (cnt == num) {
-    remove(filename);
-    vwrite_user(user, "There %s only %d mail message%s, all now deleted.\n",
-                PLTEXT_WAS(cnt), cnt, PLTEXT_S(cnt));
-    return;
-  }
-  vwrite_user(user, "%d mail message%s deleted.\n", cnt, PLTEXT_S(cnt));
-  user->read_mail = time(0) + 1;
-}
 
 
-/*
- * Show list of people your mail is from without seeing the whole lot
- */
-void
-mail_from(UR_OBJECT user)
-{
-  char w1[ARR_SIZE], line[ARR_SIZE], filename[80], *s;
-  FILE *fp;
-  int valid, cnt, tmp1, tmp2, nmail;
 
-  sprintf(filename, "%s/%s/%s.M", USERFILES, USERMAILS, user->name);
-  fp = fopen(filename, "r");
-  if (!fp) {
-    write_user(user, "You have no mail.\n");
-    return;
-  }
-  write_user(user, "\n~BB*** Mail from ***\n\n");
-  valid = 1;
-  cnt = 0;
-  fscanf(fp, "%d %d\r", &tmp1, &tmp2);
-  for (s = fgets(line, ARR_SIZE - 1, fp); s;
-       s = fgets(line, ARR_SIZE - 1, fp)) {
-    if (*s == '\n') {
-      valid = 1;
-    }
-    sscanf(s, "%s", w1);
-    if (valid && !strcmp(colour_com_strip(w1), "From:")) {
-      vwrite_user(user, "~FC%2d)~RS %s", ++cnt, remove_first(s));
-      valid = 0;
-    }
-  }
-  fclose(fp);
-  nmail = mail_sizes(user->name, 1);
-  vwrite_user(user,
-              "\nTotal of ~OL%d~RS message%s, ~OL%d~RS of which %s new.\n\n",
-              cnt, PLTEXT_S(cnt), nmail, PLTEXT_IS(nmail));
-}
 
 
 /*
@@ -1440,66 +1239,7 @@ forward_specific_mail(UR_OBJECT user)
 /***** Message boards *****/
 
 
-/*
- * Read the message board
- */
-void
-read_board(UR_OBJECT user)
-{
-  char filename[80];
-  const char *name;
-  RM_OBJECT rm;
-  int ret;
 
-  rm = NULL;
-  if (word_count < 2) {
-    rm = user->room;
-  } else {
-    if (word_count >= 3) {
-      rm = get_room(word[1]);
-      if (!rm) {
-        write_user(user, nosuchroom);
-        return;
-      }
-      ret = atoi(word[2]);
-      read_board_specific(user, rm, ret);
-      return;
-    }
-    ret = atoi(word[1]);
-    if (ret) {
-      read_board_specific(user, user->room, ret);
-      return;
-    }
-    rm = get_room(word[1]);
-    if (!rm) {
-      write_user(user, nosuchroom);
-      return;
-    }
-    if (!has_room_access(user, rm)) {
-      write_user(user,
-                 "That room is currently private, you cannot read the board.\n");
-      return;
-    }
-  }
-  vwrite_user(user, "\n~BB*** The %s message board ***\n\n", rm->name);
-  if (is_personal_room(rm)) {
-    sprintf(filename, "%s/%s/%s.B", USERFILES, USERROOMS, rm->owner);
-  } else {
-    sprintf(filename, "%s/%s.B", DATAFILES, rm->name);
-  }
-  user->filepos = 0;
-  ret = more(user, user->socket, filename);
-  if (!ret) {
-    write_user(user, "There are no messages on the board.\n\n");
-  } else if (ret == 1) {
-    user->misc_op = 2;
-  }
-  name = user->vis ? user->recap : invisname;
-  if (rm == user->room) {
-    vwrite_room_except(user->room, user, "%s~RS reads the message board.\n",
-                       name);
-  }
-}
 
 
 /*
@@ -1582,205 +1322,9 @@ read_board_specific(UR_OBJECT user, RM_OBJECT rm, int msg_number)
 }
 
 
-/*
- * Write on the message board
- */
-void
-write_board(UR_OBJECT user, char *inpstr)
-{
-  char filename[80]; /* TODO: the max filename size should be calculated */
-  const char *name;
-  char *c;
-  FILE *fp;
-  int cnt;
-
-  if (inpstr) {
-    /* FIXME: Use sentinel other JAILED */
-    if (user->muzzled != JAILED) {
-      write_user(user, "You are muzzled, you cannot write on the board.\n");
-      return;
-    }
-    if (word_count < 2) {
-#ifdef NETLINKS
-      if (user->type == REMOTE_TYPE) {
-        /* Editor will not work over netlink because all the prompts will go wrong */
-        write_user(user,
-                   "Sorry, due to software limitations remote users cannot use the line editor.\nUse the \".write <text>\" method instead.\n");
-        return;
-      }
-#endif
-      write_user(user, "\n~BB*** Writing board message ***\n\n");
-      user->misc_op = 3;
-      editor(user, NULL);
-      return;
-    }
-    strcat(inpstr, "\n");       /* XXX: risky but hopefully it will be ok */
-  } else {
-    inpstr = user->malloc_start;
-  }
-  switch (amsys->ban_swearing) {
-  case SBMAX:
-    if (contains_swearing(inpstr)) {
-      write_user(user, noswearing);
-      return;
-    }
-    break;
-  case SBMIN:
-    if (!is_personal_room(user->room)) {
-      inpstr = censor_swear_words(inpstr);
-    }
-    break;
-  case SBOFF:
-  default:
-    /* do nothing as ban_swearing is off */
-    break;
-  }
-  if (is_personal_room(user->room)) {
-    sprintf(filename, "%s/%s/%s.B", USERFILES, USERROOMS, user->room->owner);
-  } else {
-    sprintf(filename, "%s/%s.B", DATAFILES, user->room->name);
-  }
-  fp = fopen(filename, "a");
-  if (!fp) {
-    vwrite_user(user, "%s: cannot write to file.\n", syserror);
-    write_syslog(SYSLOG | ERRLOG, 0,
-                 "ERROR: Cannot open file %s to append in write_board().\n",
-                 filename);
-    return;
-  }
-  name = user->vis ? user->recap : invisname;
-  /*
-     The posting time (PT) is the time its written in machine readable form, this
-     makes it easy for this program to check the age of each message and delete
-     as appropriate in check_messages()
-   */
-#ifdef NETLINKS
-  if (user->type == REMOTE_TYPE) {
-    sprintf(text, "PT: %d\r~OLFrom: %s@%s  %s\n", (int) (time(0)), name,
-            user->netlink->service, long_date(0));
-  } else
-#endif
-  {
-    sprintf(text, "PT: %d\r~OLFrom: %s  %s\n", (int) (time(0)), name,
-            long_date(0));
-  }
-  fputs(text, fp);
-  cnt = 0;
-  for (c = inpstr; *c; ++c) {
-    putc(*c, fp);
-    if (*c == '\n') {
-      cnt = 0;
-    } else {
-      ++cnt;
-    }
-    if (cnt == 80) {
-      putc('\n', fp);
-      cnt = 0;
-    }
-  }
-  putc('\n', fp);
-  fclose(fp);
-  write_user(user, "You write the message on the board.\n");
-  vwrite_room_except(user->room, user, "%s writes a message on the board.\n",
-                     name);
-  ++user->room->mesg_cnt;
-}
 
 
-/*
- * Wipe some messages off the board
- */
-void
-wipe_board(UR_OBJECT user)
-{
-  char filename[80];
-  const char *name;
-  RM_OBJECT rm;
-  int cnt;
 
-  rm = user->room;
-  if (word_count < 2 && ((user->level >= WIZ && !is_personal_room(rm))
-                         || (is_personal_room(rm)
-                             && (is_my_room(user, rm)
-                                 || user->level >= GOD)))) {
-    write_user(user, "Usage: wipe all\n");
-    write_user(user, "       wipe <#>\n");
-    write_user(user, "       wipe to <#>\n");
-    write_user(user, "       wipe from <#> to <#>\n");
-    return;
-  } else if (word_count < 2 && ((user->level < WIZ && !is_personal_room(rm))
-                                || (is_personal_room(rm)
-                                    && !is_my_room(user, rm)
-                                    && user->level < GOD))) {
-    write_user(user, "Usage: wipe <#>\n");
-    return;
-  }
-  if (is_personal_room(rm)) {
-    if (!is_my_room(user, rm) && user->level < GOD && !check_board_wipe(user)) {
-      return;
-    } else if (get_wipe_parameters(user) == -1) {
-      return;
-    }
-  } else {
-    if (user->level < WIZ && !(check_board_wipe(user))) {
-      return;
-    } else if (get_wipe_parameters(user) == -1) {
-      return;
-    }
-  }
-  name = user->vis ? user->recap : invisname;
-  if (is_personal_room(rm)) {
-    sprintf(filename, "%s/%s/%s.B", USERFILES, USERROOMS, rm->owner);
-  } else {
-    sprintf(filename, "%s/%s.B", DATAFILES, rm->name);
-  }
-  if (!rm->mesg_cnt) {
-    write_user(user, "There are no messages on the room board.\n");
-    return;
-  }
-  if (user->wipe_from == -1) {
-    remove(filename);
-    write_user(user, "All messages deleted.\n");
-    if (user->level < GOD || user->vis) {
-      vwrite_room_except(rm, user, "%s~RS wipes the message board.\n", name);
-    }
-    write_syslog(SYSLOG, 1,
-                 "%s wiped all messages from the board in the %s.\n",
-                 user->name, rm->name);
-    rm->mesg_cnt = 0;
-    return;
-  }
-  if (user->wipe_from > rm->mesg_cnt) {
-    vwrite_user(user, "There %s only %d message%s on the board.\n",
-                PLTEXT_IS(rm->mesg_cnt), rm->mesg_cnt,
-                PLTEXT_S(rm->mesg_cnt));
-    return;
-  }
-  cnt = wipe_messages(filename, user->wipe_from, user->wipe_to, 0);
-  if (cnt == rm->mesg_cnt) {
-    remove(filename);
-    vwrite_user(user,
-                "There %s only %d message%s on the board, all now deleted.\n",
-                PLTEXT_WAS(rm->mesg_cnt), rm->mesg_cnt,
-                PLTEXT_S(rm->mesg_cnt));
-    if (user->level < GOD || user->vis) {
-      vwrite_room_except(rm, user, "%s wipes the message board.\n", name);
-    }
-    write_syslog(SYSLOG, 1,
-                 "%s wiped all messages from the board in the %s.\n",
-                 user->name, rm->name);
-    rm->mesg_cnt = 0;
-    return;
-  }
-  rm->mesg_cnt -= cnt;
-  vwrite_user(user, "%d board message%s deleted.\n", cnt, PLTEXT_S(cnt));
-  if (user->level < GOD || user->vis) {
-    vwrite_room_except(rm, user, "%s wipes some messages from the board.\n",
-                       name);
-  }
-  write_syslog(SYSLOG, 1, "%s wiped %d message%s from the board in the %s.\n",
-               user->name, cnt, PLTEXT_S(cnt), rm->name);
-}
 
 
 /*
@@ -1918,83 +1462,7 @@ board_from(UR_OBJECT user)
 }
 
 
-/*
- * Search all the boards for the words given in the list. Rooms fixed to
- * private will be ignore if the users level is less than gatecrash_level
- */
-void
-search_boards(UR_OBJECT user)
-{
-  char filename[80], line[82], buff[(MAX_LINES + 1) * 82], w1[81], *s;
-  FILE *fp;
-  RM_OBJECT rm;
-  int w, cnt, message, yes, room_given;
 
-  if (word_count < 2) {
-    write_user(user, "Usage: search <word list>\n");
-    return;
-  }
-  /* Go through rooms */
-  cnt = 0;
-  for (rm = room_first; rm; rm = rm->next) {
-    if (!has_room_access(user, rm)) {
-      continue;
-    }
-    if (is_personal_room(rm)) {
-      sprintf(filename, "%s/%s/%s.B", USERFILES, USERROOMS, rm->owner);
-    } else {
-      sprintf(filename, "%s/%s.B", DATAFILES, rm->name);
-    }
-    fp = fopen(filename, "r");
-    if (!fp) {
-      continue;
-    }
-    /* Go through file */
-    *buff = '\0';
-    yes = message = room_given = 0;
-    for (s = fgets(line, 81, fp); s; s = fgets(line, 81, fp)) {
-      if (*s == '\n') {
-        if (yes) {
-          strcat(buff, "\n");
-          write_user(user, buff);
-        }
-        message = yes = 0;
-        *buff = '\0';
-      }
-      if (!message) {
-        *w1 = '\0';
-        sscanf(s, "%s", w1);
-        if (!strcmp(w1, "PT:")) {
-          message = 1;
-          strcpy(buff, remove_first(remove_first(s)));
-        }
-      } else {
-        strcat(buff, s);
-      }
-      for (w = 1; w < word_count; ++w) {
-        if (!yes && strstr(s, word[w])) {
-          if (!room_given) {
-            vwrite_user(user, "~BB*** %s ***\n\n", rm->name);
-            room_given = 1;
-          }
-          yes = 1;
-          ++cnt;
-        }
-      }
-    }
-    if (yes) {
-      strcat(buff, "\n");
-      write_user(user, buff);
-    }
-    fclose(fp);
-  }
-  if (cnt) {
-    vwrite_user(user, "Total of %d matching message%s.\n\n", cnt,
-                PLTEXT_S(cnt));
-  } else {
-    write_user(user, "No occurences found.\n");
-  }
-}
 
 
 /*
