+++ src/netlinks.c
@@ -19,52 +19,7 @@
 #ifdef NETLINKS
 /**************************************************************************/
 
-#define ML_ENTRY(a) ML_EXPAND a
-
-/* Netlink Commands */
-#define NLC_LIST \
-  ML_ENTRY((DISCONNECT,   "DISCONNECT"  )) \
-  ML_ENTRY((TRANSFER,     "TRANS"       )) \
-  ML_ENTRY((RELEASE,      "REL"         )) \
-  ML_ENTRY((ACTION,       "ACT"         )) \
-  ML_ENTRY((GRANTED,      "GRANTED"     )) \
-  ML_ENTRY((DENIED,       "DENIED"      )) \
-  ML_ENTRY((MESSAGE,      "MSG"         )) \
-  ML_ENTRY((ENDMESSAGE,   "EMSG"        )) \
-  ML_ENTRY((PROMPT,       "PRM"         )) \
-  ML_ENTRY((VERIFICATION, "VERIFICATION")) \
-  ML_ENTRY((VERIFY,       "VERIFY"      )) \
-  ML_ENTRY((REMOVED,      "REMVD"       )) \
-  ML_ENTRY((ERROR,        "ERROR"       )) \
-  ML_ENTRY((EXISTSQUERY,  "EXISTS?"     )) \
-  ML_ENTRY((EXISTSNAK,    "EXISTS_NO"   )) \
-  ML_ENTRY((EXISTSACK,    "EXISTS_YES"  )) \
-  ML_ENTRY((MAIL,         "MAIL"        )) \
-  ML_ENTRY((ENDMAIL,      "ENDMAIL"     )) \
-  ML_ENTRY((MAILERROR,    "MAILERROR"   )) \
-  ML_ENTRY((KEEPALIVE,    "KA"          )) \
-  ML_ENTRY((RSTAT,        "RSTAT"       )) \
-  ML_ENTRY((COUNT,        NULL          ))
-
-/*
- * The most used commands have been truncated to save bandwidth, ie ACT is
- * short for action, EMSG is short for end message. Commands that do not get
- * used much i.e., VERIFICATION have been left long for readability.
- */
-
-enum nlc_value
-{
-#define ML_EXPAND(value,name) NLC_ ## value,
-  NLC_LIST
-#undef ML_EXPAND
-};
-
-static const char *const netcom[] = {
-#define ML_EXPAND(value,name) name,
-  NLC_LIST
-#undef ML_EXPAND
-};
-
+#include "netlinks.h"
 
 /*
  * Construct link object
@@ -1640,290 +1595,6 @@ mail_nl(UR_OBJECT user, char *to, const char *mesg)
   /* Rest of delivery process now up to netlink functions */
   write_user(user, "Mail sent to external talker.\n");
   return 1;
-}
-
-/******************************************************************************
- User executed commands that relate to the Netlinks
- *****************************************************************************/
-
-/*
- * Return to home site
- */
-void
-home(UR_OBJECT user)
-{
-  if (user->room) {
-    write_user(user, "You are already on your home system.\n");
-    return;
-  }
-  write_user(user, "~FB~OLYou traverse cyberspace...\n");
-  write_syslog(NETLOG, 1, "NETLINK: %s returned from %s.\n", user->name,
-               user->netlink->service);
-  release_nl(user);
-  if (user->vis) {
-    vwrite_room_except(user->room, user, "%s~RS %s\n", user->recap,
-                       user->in_phrase);
-  } else {
-    write_room_except(user->room, invisenter, user);
-  }
-  look(user);
-}
-
-
-/*
- * List defined netlinks and their status
- */
-void
-netstat(UR_OBJECT user)
-{
-  static const char *const allow[] = { "  ?", "ALL", " IN", "OUT" };
-  static const char *const type[] = { "  -", " IN", "OUT" };
-  char nlstat[9], vers[8];
-  NL_OBJECT nl;
-  UR_OBJECT u;
-  int iu, ou, a;
-
-  if (!nl_first) {
-    write_user(user, "No remote connections configured.\n");
-    return;
-  }
-  write_user(user,
-             "\n~BB*** Netlink data & status ***\n\n~FCService name    : Allow Type Status IU OU Version  Site\n\n");
-  for (nl = nl_first; nl; nl = nl->next) {
-    iu = ou = 0;
-    if (nl->stage == UP) {
-      for (u = user_first; u; u = u->next) {
-        if (u->netlink == nl) {
-          if (u->type == REMOTE_TYPE) {
-            ++iu;
-          }
-          if (!u->room) {
-            ++ou;
-          }
-        }
-      }
-    }
-    if (nl->type == UNCONNECTED) {
-      strcpy(nlstat, "~FRDOWN");
-      strcpy(vers, "-");
-    } else {
-      if (nl->stage == UP) {
-        strcpy(nlstat, "  ~FGUP");
-      } else {
-        strcpy(nlstat, " ~FYVER");
-      }
-      if (!nl->ver_major) {
-        /* Pre - 3.2 version */
-        strcpy(vers, "3.?.?");
-      } else {
-        sprintf(vers, "%d.%d.%d", nl->ver_major, nl->ver_minor,
-                nl->ver_patch);
-      }
-    }
-    /*
-       If link is incoming and remoter vers < 3.2 we have no way of knowing
-       what the permissions on it are so set to blank
-     */
-    if (!nl->ver_major && nl->type == INCOMING && nl->allow != IN) {
-      a = 0;
-    } else {
-      a = nl->allow + 1;
-    }
-    sprintf(text, "%-15s :   %s  %s   %s~RS %2d %2d %7s  %s %s\n",
-            nl->service, allow[a], type[nl->type], nlstat, iu, ou, vers,
-            nl->site, nl->port);
-    write_user(user, text);
-  }
-  write_user(user, "\n");
-}
-
-
-/*
- * Show type of data being received down links (this is usefull when a link has hung)
- */
-void
-netdata(UR_OBJECT user)
-{
-  char from[80], name[USER_NAME_LEN + 1];
-  NL_OBJECT nl;
-  int cnt;
-
-  cnt = 0;
-  write_user(user, "\n~BB*** Mail receiving status ***\n\n");
-  for (nl = nl_first; nl; nl = nl->next) {
-    if (nl->type == UNCONNECTED || !nl->mailfile) {
-      continue;
-    }
-    if (!cnt++) {
-      write_user(user,
-                 "To              : From                       Last recv.\n\n");
-    }
-    sprintf(from, "%s@%s", nl->mail_from, nl->service);
-    sprintf(text, "%-15s : %-25s  %d seconds ago.\n", nl->mail_to, from,
-            (int) (time(0) - nl->last_recvd));
-    write_user(user, text);
-  }
-  if (!cnt) {
-    write_user(user, "No mail being received.\n\n");
-  } else {
-    write_user(user, "\n");
-  }
-  cnt = 0;
-  write_user(user, "\n~BB*** Message receiving status ***\n\n");
-  for (nl = nl_first; nl; nl = nl->next) {
-    if (nl->type == UNCONNECTED || !nl->mesg_user) {
-      continue;
-    }
-    if (!cnt++) {
-      write_user(user, "To              : From             Last recv.\n\n");
-    }
-    if (nl->mesg_user == (UR_OBJECT) -1) {
-      strcpy(name, "<unknown>");
-    } else {
-      strcpy(name, nl->mesg_user->name);
-    }
-    sprintf(text, "%-15s : %-15s  %d seconds ago.\n", name, nl->service,
-            (int) (time(0) - nl->last_recvd));
-    write_user(user, text);
-  }
-  if (!cnt) {
-    write_user(user, "No messages being received.\n\n");
-  } else {
-    write_user(user, "\n");
-  }
-}
-
-
-/*
- * Connect a netlink. Use the room as the key
- */
-void
-connect_netlink(UR_OBJECT user)
-{
-  RM_OBJECT rm;
-  NL_OBJECT nl;
-
-  if (word_count < 2) {
-    write_user(user, "Usage: connect <room service is linked to>\n");
-    return;
-  }
-  rm = get_room(word[1]);
-  if (!rm) {
-    write_user(user, nosuchroom);
-    return;
-  }
-  nl = rm->netlink;
-  if (!nl) {
-    write_user(user, "That room is not linked to a service.\n");
-    return;
-  }
-  if (nl->type != UNCONNECTED) {
-    write_user(user, "That rooms netlink is already up.\n");
-    return;
-  }
-  write_user(user,
-             "Attempting connect (this may cause a temporary hang)...\n");
-  write_syslog(NETLOG, 1,
-               "NETLINK: Connection attempt to %s initiated by %s.\n",
-               nl->service, user->name);
-  nl->socket = socket_connect(nl->site, nl->port);
-  if (nl->socket < 0) {
-    vwrite_user(user, "~FRConnect failed: %s.\n",
-                nl->socket != -1 ? "Unknown hostname" : strerror(errno));
-    write_syslog(NETLOG, 1, "NETLINK: Failed to connect to %s: %s.\n",
-                 nl->service,
-                 nl->socket != -1 ? "Unknown hostname" : strerror(errno));
-    return;
-  }
-  nl->type = OUTGOING;
-  nl->stage = VERIFYING;
-  nl->last_recvd = time(0);
-  nl->connect_room = rm;
-  write_user(user, "~FGInitial connection made...\n");
-  write_syslog(NETLOG, 1, "NETLINK: Connected to %s (%s:%s).\n", nl->service,
-               nl->site, nl->port);
-}
-
-
-/*
- * Disconnect a link
- */
-void
-disconnect_netlink(UR_OBJECT user)
-{
-  RM_OBJECT rm;
-  NL_OBJECT nl;
-
-  if (word_count < 2) {
-    write_user(user, "Usage: disconnect <room service is linked to>\n");
-    return;
-  }
-  rm = get_room(word[1]);
-  if (!rm) {
-    write_user(user, nosuchroom);
-    return;
-  }
-  nl = rm->netlink;
-  if (!nl) {
-    write_user(user, "That room is not linked to a service.\n");
-    return;
-  }
-  if (nl->type == UNCONNECTED) {
-    write_user(user, "That rooms netlink is not connected.\n");
-    return;
-  }
-  /* If link has hung at verification stage do not bother announcing it */
-  if (nl->stage == UP) {
-    sprintf(text, "~OLSYSTEM:~RS Disconnecting from %s in the %s.\n",
-            nl->service, rm->name);
-    write_room(NULL, text);
-    write_syslog(NETLOG, 1,
-                 "NETLINK: Link to %s in the %s disconnected by %s.\n",
-                 nl->service, rm->name, user->name);
-  } else {
-    write_syslog(NETLOG, 1, "NETLINK: Link to %s disconnected by %s.\n",
-                 nl->service, user->name);
-  }
-  shutdown_netlink(nl);
-  write_user(user, "Disconnected.\n");
-}
-
-
-/*
- * Stat a remote system
- */
-void
-remote_stat(UR_OBJECT user)
-{
-  NL_OBJECT nl;
-  RM_OBJECT rm;
-
-  if (word_count < 2) {
-    write_user(user, "Usage: rstat <room service is linked to>\n");
-    return;
-  }
-  rm = get_room(word[1]);
-  if (!rm) {
-    write_user(user, nosuchroom);
-    return;
-  }
-  nl = rm->netlink;
-  if (!nl) {
-    write_user(user, "That room is not linked to a service.\n");
-    return;
-  }
-  if (nl->stage != UP) {
-    write_user(user, "Not (fully) connected to service.\n");
-    return;
-  }
-  if (nl->ver_major <= 3 && nl->ver_minor < 1) {
-    write_user(user,
-               "The NUTS version running that service does not support this facility.\n");
-    return;
-  }
-  sprintf(text, "%s %s\n", netcom[NLC_RSTAT], user->name);
-  write_sock(nl->socket, text);
-  write_user(user, "Request sent.\n");
 }
 
 #else
