+++ src/rooms.c
@@ -14,23 +14,7 @@
 #include "globals.h"
 #include "commands.h"
 #include "prototypes.h"
-
-struct priv_room_struct
-{
-  const char *name;
-  enum lvl_value level;
-};
-
-/*
- * The rooms listed here are just examples of what can be added
- * You may add more or remove as many as you like, but you MUST
- * keep the stopping clause in
- */
-static const struct priv_room_struct priv_room[] = {
-  {"wizroom", WIZ},             /* a room for wizzes+ only */
-  {"andys_computer", GOD},      /* only top people can get in this place! */
-  {NULL, NUM_LEVELS}            /* stopping clause */
-};
+#include "rooms.h"
 
 /***************************************************************************/
 
@@ -831,69 +815,6 @@ personal_room_rename(UR_OBJECT user, char *inpstr)
 }
 
 
-/*
- * Move to another room
- */
-void
-go(UR_OBJECT user)
-{
-  RM_OBJECT rm;
-  int i;
-
-  if (user->lroom == 2) {
-    write_user(user, "You have been shackled and cannot move.\n");
-    return;
-  }
-  if (word_count < 2) {
-    rm = get_room_full(amsys->default_warp);
-    if (!rm) {
-      write_user(user, "You cannot warp to the main room at this time.\n");
-      return;
-    }
-    if (user->room == rm) {
-      vwrite_user(user, "You are already in the %s!\n", rm->name);
-      return;
-    }
-    move_user(user, rm, 1);
-    return;
-  }
-#ifdef NETLINKS
-  release_nl(user);
-  if (transfer_nl(user)) {
-    return;
-  }
-#endif
-  rm = get_room(word[1]);
-  if (!rm) {
-    write_user(user, nosuchroom);
-    return;
-  }
-  if (rm == user->room) {
-    vwrite_user(user, "You are already in the %s!\n", rm->name);
-    return;
-  }
-  /* See if link from current room */
-  for (i = 0; i < MAX_LINKS; ++i) {
-    if (user->room->link[i] == rm) {
-      break;
-    }
-  }
-  if (i < MAX_LINKS) {
-    move_user(user, rm, 0);
-    return;
-  }
-  if (is_personal_room(rm)) {
-    write_user(user,
-               "To go to another user's home you must \".visit\" them.\n");
-    return;
-  }
-  if (user->level < WIZ) {
-    vwrite_user(user, "The %s is not adjoined to here.\n", rm->name);
-    return;
-  }
-  move_user(user, rm, 1);
-}
-
 
 /*
  * Called by go() and move()
@@ -953,155 +874,7 @@ move_user(UR_OBJECT user, RM_OBJECT rm, int teleport)
 }
 
 
-/*
- * Wizard moves a user to another room
- */
-void
-move(UR_OBJECT user)
-{
-  UR_OBJECT u;
-  RM_OBJECT rm;
-  const char *name;
 
-  if (word_count < 2) {
-    write_user(user, "Usage: move <user> [<room>]\n");
-    return;
-  }
-  u = get_user_name(user, word[1]);
-  if (!u) {
-    write_user(user, notloggedon);
-    return;
-  }
-  if (u->login) {
-    write_user(user, "That user is not logged in.\n");
-    return;
-  }
-  if (word_count < 3) {
-    rm = user->room;
-  } else {
-    rm = get_room(word[2]);
-    if (!rm) {
-      write_user(user, nosuchroom);
-      return;
-    }
-  }
-  if (user == u) {
-    write_user(user,
-               "Trying to move yourself this way is the fourth sign of madness.\n");
-    return;
-  }
-  if (u->level >= user->level) {
-    write_user(user,
-               "You cannot move a user of equal or higher level than yourself.\n");
-    return;
-  }
-  if (rm == u->room) {
-    vwrite_user(user, "%s~RS is already in the %s.\n", u->recap, rm->name);
-    return;
-  };
-  if (!has_room_access(user, rm)) {
-    vwrite_user(user,
-                "The %s is currently private, %s~RS cannot be moved there.\n",
-                rm->name, u->recap);
-    return;
-  }
-  write_user(user, "~FC~OLYou chant an ancient spell...\n");
-  name = user->vis ? user->recap : invisname;
-  if (!user->vis) {
-    write_monitor(user, user->room, 0);
-  }
-  vwrite_room_except(user->room, user,
-                     "%s~RS ~FC~OLchants an ancient spell...\n", name);
-  move_user(u, rm, 2);
-  prompt(u);
-}
-
-
-/*
- * Set rooms to public or private
- */
-void
-set_room_access(UR_OBJECT user, int priv)
-{
-  UR_OBJECT u;
-  RM_OBJECT rm;
-  const char *name;
-
-  if (word_count < 2) {
-    rm = user->room;
-  } else {
-    if (user->level < amsys->gatecrash_level) {
-      write_user(user,
-                 "You are not a high enough level to use the room option.\n");
-      return;
-    }
-    rm = get_room(word[1]);
-    if (!rm) {
-      write_user(user, nosuchroom);
-      return;
-    }
-  }
-  if (is_personal_room(rm)) {
-    if (rm == user->room) {
-      write_user(user, "This room's access is personal.\n");
-    } else {
-      write_user(user, "That room's access is personal.\n");
-    }
-    return;
-  }
-  if (is_fixed_room(rm)) {
-    if (rm == user->room) {
-      write_user(user, "This room's access is fixed.\n");
-    } else {
-      write_user(user, "That room's access is fixed.\n");
-    }
-    return;
-  }
-  if (priv) {
-    if (is_private_room(rm)) {
-      if (rm == user->room) {
-        write_user(user, "This room is already private.\n");
-      } else {
-        write_user(user, "That room is already private.\n");
-      }
-      return;
-    }
-    if (room_visitor_count(rm) < amsys->min_private_users
-        && user->level < amsys->ignore_mp_level) {
-      vwrite_user(user,
-                  "You need at least %d users/clones in a room before it can be made private.\n",
-                  amsys->min_private_users);
-      return;
-    }
-  } else {
-    if (!is_private_room(rm)) {
-      if (rm == user->room) {
-        write_user(user, "This room is already public.\n");
-      } else {
-        write_user(user, "That room is already public.\n");
-      }
-      return;
-    }
-    /* Reset any invites into the room & clear review buffer */
-    for (u = user_first; u; u = u->next) {
-      if (u->invite_room == rm) {
-        u->invite_room = NULL;
-      }
-    }
-    clear_revbuff(rm);
-  }
-  rm->access ^= PRIVATE;
-  name = user->vis ? user->recap : invisname;
-  if (rm == user->room) {
-    vwrite_room_except(rm, user, "%s~RS has set the room to %s~RS.\n", name,
-                       is_private_room(rm) ? "~FRPRIVATE" : "~FGPUBLIC");
-  } else {
-    vwrite_room(rm, "This room has been set to %s~RS.\n",
-                is_private_room(rm) ? "~FRPRIVATE" : "~FGPUBLIC");
-  }
-  vwrite_user(user, "Room set to %s~RS.\n",
-              is_private_room(rm) ? "~FRPRIVATE" : "~FGPUBLIC");
-}
 
 
 /*
@@ -1158,50 +931,7 @@ change_room_fix(UR_OBJECT user, int fix)
 }
 
 
-/*
- * Invite a user into a private room
- */
-void
-invite(UR_OBJECT user)
-{
-  UR_OBJECT u;
-  RM_OBJECT rm;
-  const char *name;
 
-  if (word_count < 2) {
-    write_user(user, "Invite who?\n");
-    return;
-  }
-  rm = user->room;
-  if (!is_private_room(rm)) {
-    write_user(user, "This room is currently public.\n");
-    return;
-  }
-  u = get_user_name(user, word[1]);
-  if (!u) {
-    write_user(user, notloggedon);
-    return;
-  }
-  if (u == user) {
-    write_user(user,
-               "Inviting yourself to somewhere is the third sign of madness.\n");
-    return;
-  }
-  if (u->room == rm) {
-    vwrite_user(user, "%s~RS is already here!\n", u->recap);
-    return;
-  }
-  if (u->invite_room == rm) {
-    vwrite_user(user, "%s~RS has already been invited into here.\n",
-                u->recap);
-    return;
-  }
-  vwrite_user(user, "You invite %s~RS in.\n", u->recap);
-  name = user->vis || u->level >= user->level ? user->recap : invisname;
-  vwrite_user(u, "%s~RS has invited you into the %s.\n", name, rm->name);
-  u->invite_room = rm;
-  strcpy(u->invite_by, user->name);
-}
 
 
 /*
@@ -1240,203 +970,6 @@ uninvite(UR_OBJECT user)
   vwrite_user(u, "%s~RS cancels your invitation.\n", name);
   u->invite_room = NULL;
   *u->invite_by = '\0';
-}
-
-
-/*
- * Ask to be let into a private room
- */
-void
-letmein(UR_OBJECT user)
-{
-  RM_OBJECT rm;
-  int i;
-
-  if (word_count < 2) {
-    write_user(user, "Knock on what door?\n");
-    return;
-  }
-  rm = get_room(word[1]);
-  if (!rm) {
-    write_user(user, nosuchroom);
-    return;
-  }
-  if (rm == user->room) {
-    vwrite_user(user, "You are already in the %s!\n", rm->name);
-    return;
-  }
-  for (i = 0; i < MAX_LINKS; ++i)
-    if (user->room->link[i] == rm) {
-      break;
-    }
-  if (i >= MAX_LINKS) {
-    vwrite_user(user, "The %s is not adjoined to here.\n", rm->name);
-    return;
-  }
-  if (!is_private_room(rm)) {
-    vwrite_user(user, "The %s is currently public.\n", rm->name);
-    return;
-  }
-  vwrite_user(user, "You knock asking to be let into the %s.\n", rm->name);
-  vwrite_room_except(user->room, user,
-                     "%s~RS knocks asking to be let into the %s.\n",
-                     user->recap, rm->name);
-  vwrite_room(rm, "%s~RS knocks asking to be let in.\n", user->recap);
-}
-
-
-/*
- * Show talker rooms
- */
-void
-rooms(UR_OBJECT user, int show_topics, int wrap)
-{
-  RM_OBJECT rm;
-  UR_OBJECT u;
-#ifdef NETLINKS
-  NL_OBJECT nl;
-  char serv[SERV_NAME_LEN + 1], nstat[9];
-  char rmaccess[9];
-#endif
-  int cnt, rm_cnt, rm_pub, rm_priv;
-
-  if (word_count < 2) {
-    if (!wrap) {
-      user->wrap_room = room_first;
-    }
-    if (show_topics) {
-      write_user(user,
-                 "\n+----------------------------------------------------------------------------+\n");
-      write_user(user,
-                 "~FC~OLpl  u/m~RS  | ~OL~FCname~RS                 - ~FC~OLtopic\n");
-      write_user(user,
-                 "+----------------------------------------------------------------------------+\n");
-    } else {
-      write_user(user,
-                 "\n+----------------------------------------------------------------------------+\n");
-      write_user(user,
-                 "~FC~OLRoom name            ~RS|~FC~OL Access  Users  Mesgs  Inlink  LStat  Service\n");
-      write_user(user,
-                 "+----------------------------------------------------------------------------+\n");
-    }
-    rm_cnt = 0;
-    for (rm = user->wrap_room; rm; rm = rm->next) {
-      if (is_personal_room(rm)) {
-        continue;
-      }
-      if (rm_cnt == user->pager - 4) {
-        switch (show_topics) {
-        case 0:
-          user->misc_op = 10;
-          break;
-        case 1:
-          user->misc_op = 11;
-          break;
-        }
-        write_user(user, "~BB~FG-=[*]=- PRESS <RETURN>, E TO EXIT:~RS ");
-        return;
-      }
-      cnt = 0;
-      for (u = user_first; u; u = u->next)
-        if (u->type != CLONE_TYPE && u->room == rm) {
-          ++cnt;
-        }
-      if (show_topics) {
-        vwrite_user(user, "%c%c %2d/%-2d | %s%-20.20s~RS - %s\n",
-                    is_private_room(rm) ? 'P' : ' ',
-                    is_fixed_room(rm) ? '*' : ' ', cnt, rm->mesg_cnt,
-                    is_private_room(rm) ? "~FR~OL" : "", rm->name, rm->topic);
-      }
-#ifdef NETLINKS
-      else {
-        if (is_private_room(rm)) {
-          strcpy(rmaccess, " ~FRPRIV");
-        } else {
-          strcpy(rmaccess, " ~FGPUB ");
-        }
-        if (is_fixed_room(rm)) {
-          *rmaccess = '*';
-        }
-        nl = rm->netlink;
-        *serv = '\0';
-        if (!nl) {
-          if (rm->inlink) {
-            strcpy(nstat, " ~FRDOWN");
-          } else {
-            strcpy(nstat, "    -");
-          }
-        } else {
-          if (nl->type == UNCONNECTED) {
-            strcpy(nstat, " ~FRDOWN");
-          } else if (nl->stage == UP) {
-            strcpy(nstat, "   ~FGUP");
-          } else {
-            strcpy(nstat, "  ~FYVER");
-          }
-        }
-        if (nl) {
-          strcpy(serv, nl->service);
-        }
-        vwrite_user(user, "%-20s | %9s~RS  %5d  %5d  %-6s  %s~RS  %s\n",
-                    rm->name, rmaccess, cnt, rm->mesg_cnt, noyes[rm->inlink],
-                    nstat, serv);
-      }
-#endif
-      ++rm_cnt;
-      user->wrap_room = rm->next;
-    }
-    user->misc_op = 0;
-    rm_pub = rm_priv = 0;
-    for (rm = room_first; rm; rm = rm->next) {
-      if (is_personal_room(rm)) {
-        continue;
-      }
-      if (is_private_room(rm)) {
-        ++rm_priv;
-      } else {
-        ++rm_pub;
-      }
-    }
-    write_user(user,
-               "+----------------------------------------------------------------------------+\n");
-    vwrite_user(user,
-                "There is a total of ~OL%d~RS rooms.  ~OL%d~RS %s public, and ~OL%d~RS %s private.\n",
-                rm_priv + rm_pub, rm_pub, PLTEXT_IS(rm_pub), rm_priv,
-                PLTEXT_IS(rm_priv));
-    write_user(user,
-               "+----------------------------------------------------------------------------+\n\n");
-    return;
-  }
-  if (!strcasecmp(word[1], "-l")) {
-    write_user(user, "The following rooms are default...\n\n");
-    vwrite_user(user, "Default main room : ~OL%s\n", room_first->name);
-    vwrite_user(user, "Default warp room : ~OL%s\n",
-                *amsys->default_warp ? amsys->default_warp : "<none>");
-    vwrite_user(user, "Default jail room : ~OL%s\n",
-                *amsys->default_jail ? amsys->default_jail : "<none>");
-#ifdef GAMES
-    vwrite_user(user, "Default bank room : ~OL%s\n",
-                *amsys->default_bank ? amsys->default_bank : "<none>");
-    vwrite_user(user, "Default shoot room : ~OL%s\n",
-                *amsys->default_shoot ? amsys->default_shoot : "<none>");
-#endif
-    if (!priv_room[0].name) {
-      write_user(user,
-                 "\nThere are no level specific rooms currently availiable.\n\n");
-      return;
-    }
-    write_user(user, "\nThe following rooms are level specific...\n\n");
-    for (cnt = 0; priv_room[cnt].name; ++cnt) {
-      vwrite_user(user,
-                  "~FC%s~RS is for users of level ~OL%s~RS and above.\n",
-                  priv_room[cnt].name, user_level[priv_room[cnt].level].name);
-    }
-    vwrite_user(user,
-                "\nThere is a total of ~OL%d~RS level specific rooms.\n\n",
-                cnt);
-    return;
-  }
-  write_user(user, "Usage: rooms [-l]\n");
 }
 
 
@@ -1527,51 +1060,6 @@ join(UR_OBJECT user)
 
 
 
-/*
- * Set the room topic
- */
-void
-set_topic(UR_OBJECT user, char *inpstr)
-{
-  RM_OBJECT rm;
-  const char *name;
-
-  rm = user->room;
-  if (word_count < 2) {
-    if (!*rm->topic) {
-      write_user(user, "No topic has been set yet.\n");
-      return;
-    }
-    vwrite_user(user, "The current topic is: %s\n", rm->topic);
-    return;
-  }
-  if (strlen(inpstr) > TOPIC_LEN) {
-    write_user(user, "Topic too long.\n");
-    return;
-  }
-  switch (amsys->ban_swearing) {
-  case SBMAX:
-    if (contains_swearing(inpstr)) {
-      write_user(user, noswearing);
-      return;
-    }
-    break;
-  case SBMIN:
-    if (!is_personal_room(user->room)) {
-      inpstr = censor_swear_words(inpstr);
-    }
-    break;
-  case SBOFF:
-  default:
-    /* do nothing as ban_swearing is off */
-    break;
-  }
-  vwrite_user(user, "Topic set to: %s\n", inpstr);
-  name = user->vis ? user->recap : invisname;
-  vwrite_room_except(rm, user, "%s~RS has set the topic to: %s\n", name,
-                     inpstr);
-  strcpy(rm->topic, inpstr);
-}
